Added modularity to the code
BubaCoin became 'Contract', CryoCoin now is not useful anymore because you can call the constructor of Contract and pass directly into it the abi and the address of the contract. Contract is a class that has the basic function of every ERC20 token, so we have 1 class for all of our three tokens. There's one problem. The functions return undefinied because they return a promise, I don't know how to fix it yet but the key to solve this is here (https://stackoverflow.com/questions/52688855/how-to-fix-async-function-which-returns-undefined).